implement go2cpp and cpp2go
