+++ golang/README.md
@@ -11,13 +11,14 @@ To run integration examples :
 
 Running vanilla-rtb bidder as applcation and linked in library written in Go as a bid handler 
 ```
-npm install
+BUILDTYPE=cpp2go npm install
 ./build/vanilla-rtb-go --config config.cfg 
 ```
 
 This process builds bid_handler.a from bid_handler.go and links it with bidder.cpp and other vanilla-rtb sources 
 
-The process utlizes following commands 
+
+The process utlizes following commands  
 
 ```
 go run  bidder_generator.go --output-dir golang/ --input-template templates/bidder.tmpl -g app -T ico -B APP 
@@ -27,21 +28,26 @@ go build -buildmode=c-archive bid_handler.go
 
 ```
 
-Another way ( still work in progress ) is to use bid_handler.go not only as a handler but also as a main entry point 
-and have bidder.cpp generted and compiled as a library ( e.g. __main__() instead of main() generated with -B LIB )
+Another way is to use bidder.go not only  as a handler but also as a main entry point 
+```
+BUILDTYPE=go2cpp npm install
+./bidder --config config.cfg 
+```
 
+What happens in this case is bidder.cpp is generted and compiled as a library ( e.g. __main__() instead of main() generated with -B LIB )
+Internally preinstall.sh and CMakelists.txt are using following commands
 ```
 go run  bidder_generator.go --output-dir golang/ --input-template templates/bidder.tmpl -g app -T ico -B LIB
 go run  bidder_generator.go --output-dir golang/ --input-template templates/matcher.tmpl -g matchers
 cd golang
-go build -buildmode=c-archive bid_handler.go
-export CGO_LDFLAGS="-L build bid_handler.a vanilla-rtb-go.a  -L/path/to/boost/lib -lboost_program_options -lboost_system -lboost_regex" 
-go build -buildmode=exe bid_handler.go
+go build -buildmode=c-archive bidder.go
+export CGO_LDFLAGS="bidder.a vanilla-rtb-go.a"
+go build -buildmode=exe bidder.go
 ```
 
-Running second type of build is slightly different as the last command will link everything into bid_handler executable 
+Running second type of build is slightly different as the last command will link everything into bidder executable 
 
 ```
-./bid_handler --config config.cfg 
+./bidder --config config.cfg 
 ```
